+++ test/integrationtest/projecttest/modules/cart/module.go
@@ -16,6 +16,7 @@ type (
 func (m *Module) Configure(injector *dingo.Injector) {
 	injector.Override((*infrastructure.VoucherHandler)(nil), "").To(&FakeVoucherHandler{})
 	injector.BindMulti((*validation.MaxQuantityRestrictor)(nil)).To(FakeQtyRestrictor{})
+	injector.Bind(new(validation.PaymentSelectionValidator)).To(new(FakePaymentSelectionValidator))
 }
 
 // Depends on other modules